backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Axel791 <HS.Kazakov@mail.ru>"]
license = "BSD-3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.5"
django-filter = "^23.2"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
pillow = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
